照準の色を変える処理の追加（仮）

+++ k2Engine-master/GameTemplate/Game/AnnoyingEnemy.h
@@ -102,13 +102,13 @@ public:
 	}
 
 	//HPを取得する関数。
-	float GetHP()const
+	int GetHP()const
 	{
 		return	m_enemyHP;
 	}
 
 	//最大HPを取得する関数。
-	float GetMaxHP()const
+	int GetMaxHP()const
 	{
 		return m_enemyMaxHP;
 	}
